moved all other mixinx into the new minecraft package, cleaned them up and added missing annotations
