+++ src/main/java/mod/beethoven92/betterendforge/mixin/minecraft/BoneMealItemMixin.java
@@ -1,4 +1,4 @@
-package mod.beethoven92.betterendforge.mixin;
+package mod.beethoven92.betterendforge.mixin.minecraft;
 
 import java.util.Random;
 
@@ -13,6 +13,7 @@ import net.minecraft.util.EnumActionResult;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.EnumHand;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -29,15 +30,19 @@ import net.minecraft.world.World;
 // TO DO: consider replacing the mixin with events and block behaviours(see grass block grow method for example)
 @Mixin(ItemDye.class)
 public abstract class BoneMealItemMixin {
-    private static final EnumFacing[] DIR = BlockHelper.makeHorizontal();
-    private static final Mutable POS = new Mutable();
+    @Unique
+    private static final EnumFacing[] betterEndForge_1_12_2$DIR = BlockHelper.makeHorizontal();
+    @Unique
+    private static final Mutable betterEndForge_1_12_2$POS = new Mutable();
 
     @Inject(method = "onItemUse", at = @At("HEAD"), cancellable = true)
     private void onItemUse(EntityPlayer player, World world, BlockPos blockPos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ, CallbackInfoReturnable<EnumActionResult> info) {
         ItemStack itemstack = player.getHeldItem(hand);
         EnumDyeColor enumdyecolor = EnumDyeColor.byDyeDamage(itemstack.getMetadata());
 
-        if (enumdyecolor != EnumDyeColor.WHITE) return;
+        if (enumdyecolor != EnumDyeColor.WHITE) {
+            return;
+        }
         // FIX underwater seeds not being able to grow when using bonemeal on them
         if (ItemDye.applyBonemeal(itemstack, world, blockPos, player, hand)) {
             if (!world.isRemote) {
@@ -53,7 +58,7 @@ public abstract class BoneMealItemMixin {
             if (ModTags.END_GROUND.contains(world.getBlockState(blockPos).getBlock())) {
                 boolean consume = false;
                 if (world.getBlockState(blockPos).getBlock() == (Blocks.END_STONE)) {
-                    IBlockState nylium = getNylium(world, blockPos);
+                    IBlockState nylium = betterEndForge_1_12_2$getNylium(world, blockPos);
                     if (nylium != null) {
                         BlockHelper.setWithoutUpdate(world, blockPos, nylium);
                         consume = true;
@@ -66,15 +71,18 @@ public abstract class BoneMealItemMixin {
                 } else {
                     if (world.getBlockState(offseted).getBlock() == Blocks.WATER && endBiome) {
                         // FIX being able to use bone meal underwater on a block where plants already grew
-                        if (world.getBlockState(offseted).getBlock() == (Blocks.WATER))
-                            consume = growWaterGrass(world, blockPos);
+                        if (world.getBlockState(offseted).getBlock() == (Blocks.WATER)) {
+                            consume = betterEndForge_1_12_2$growWaterGrass(world, blockPos);
+                        }
                     } else if (world.getBlockState(offseted).getBlock() == Blocks.AIR) // FIX being able to use bone meal on a block where plants already grew
                     {
-                        consume = growGrass(world, blockPos);
+                        consume = betterEndForge_1_12_2$growGrass(world, blockPos);
                     }
                 }
                 if (consume) {
-                    if (!player.isCreative()) itemstack.shrink(1);
+                    if (!player.isCreative()) {
+                        itemstack.shrink(1);
+                    }
                     world.playEvent(2005, blockPos, 0);
                     info.setReturnValue(EnumActionResult.SUCCESS);
                     info.cancel();
@@ -88,22 +96,23 @@ public abstract class BoneMealItemMixin {
         }
     }
 
-    private boolean growGrass(World world, BlockPos pos) {
+    @Unique
+    private boolean betterEndForge_1_12_2$growGrass(World world, BlockPos pos) {
         int y1 = pos.getY() + 3;
         int y2 = pos.getY() - 3;
         boolean result = false;
         for (int i = 0; i < 64; i++) {
             int x = (int) (pos.getX() + world.rand.nextGaussian() * 2);
             int z = (int) (pos.getZ() + world.rand.nextGaussian() * 2);
-            POS.setX(x);
-            POS.setZ(z);
+            betterEndForge_1_12_2$POS.setX(x);
+            betterEndForge_1_12_2$POS.setZ(z);
             for (int y = y1; y >= y2; y--) {
-                POS.setY(y);
-                BlockPos down = POS.down();
-                if (world.isAirBlock(POS) && !world.isAirBlock(down)) {
-                    IBlockState grass = getGrassState(world, down);
+                betterEndForge_1_12_2$POS.setY(y);
+                BlockPos down = betterEndForge_1_12_2$POS.down();
+                if (world.isAirBlock(betterEndForge_1_12_2$POS) && !world.isAirBlock(down)) {
+                    IBlockState grass = betterEndForge_1_12_2$getGrassState(world, down);
                     if (grass != null) {
-                        BlockHelper.setWithoutUpdate(world, POS, grass);
+                        BlockHelper.setWithoutUpdate(world, betterEndForge_1_12_2$POS, grass);
                         result = true;
                     }
                     break;
@@ -113,22 +122,23 @@ public abstract class BoneMealItemMixin {
         return result;
     }
 
-    private boolean growWaterGrass(World world, BlockPos pos) {
+    @Unique
+    private boolean betterEndForge_1_12_2$growWaterGrass(World world, BlockPos pos) {
         int y1 = pos.getY() + 3;
         int y2 = pos.getY() - 3;
         boolean result = false;
         for (int i = 0; i < 64; i++) {
             int x = (int) (pos.getX() + world.rand.nextGaussian() * 2);
             int z = (int) (pos.getZ() + world.rand.nextGaussian() * 2);
-            POS.setX(x);
-            POS.setZ(z);
+            betterEndForge_1_12_2$POS.setX(x);
+            betterEndForge_1_12_2$POS.setZ(z);
             for (int y = y1; y >= y2; y--) {
-                POS.setY(y);
-                BlockPos down = POS.down();
-                if (world.getBlockState(POS).getBlock() == (Blocks.WATER) && ModTags.END_GROUND.contains(world.getBlockState(down).getBlock())) {
-                    IBlockState grass = getWaterGrassState(world, down);
+                betterEndForge_1_12_2$POS.setY(y);
+                BlockPos down = betterEndForge_1_12_2$POS.down();
+                if (world.getBlockState(betterEndForge_1_12_2$POS).getBlock() == (Blocks.WATER) && ModTags.END_GROUND.contains(world.getBlockState(down).getBlock())) {
+                    IBlockState grass = betterEndForge_1_12_2$getWaterGrassState(world, down);
                     if (grass != null) {
-                        BlockHelper.setWithoutUpdate(world, POS, grass);
+                        BlockHelper.setWithoutUpdate(world, betterEndForge_1_12_2$POS, grass);
                         result = true;
                     }
                     break;
@@ -138,18 +148,20 @@ public abstract class BoneMealItemMixin {
         return result;
     }
 
-    private Block[] glowingGrasslandsGrass() {
+    @Unique
+    private Block[] betterEndForge_1_12_2$glowingGrasslandsGrass() {
         return new Block[]{ModBlocks.BLOOMING_COOKSONIA, ModBlocks.VAIOLUSH_FERN,
                 ModBlocks.FRACTURN, ModBlocks.SALTEAGO, ModBlocks.CREEPING_MOSS,
                 ModBlocks.UMBRELLA_MOSS, ModBlocks.TWISTED_UMBRELLA_MOSS};
     }
 
-    private IBlockState getGrassState(World world, BlockPos pos) {
+    @Unique
+    private IBlockState betterEndForge_1_12_2$getGrassState(World world, BlockPos pos) {
         IBlockState state = world.getBlockState(pos);
         Block block = state.getBlock();
         if (block == ModBlocks.END_MOSS || block == ModBlocks.END_MYCELIUM) {
             if (world.getBiome(pos).getRegistryName().equals(ModBiomes.GLOWING_GRASSLANDS.getID())) {
-                Block[] grasses = glowingGrasslandsGrass();
+                Block[] grasses = betterEndForge_1_12_2$glowingGrasslandsGrass();
                 return grasses[world.rand.nextInt(grasses.length)].getDefaultState();
             } else {
                 return world.rand.nextBoolean() ? ModBlocks.CREEPING_MOSS.getDefaultState() : ModBlocks.UMBRELLA_MOSS.getDefaultState();
@@ -167,19 +179,20 @@ public abstract class BoneMealItemMixin {
         } else if (block == ModBlocks.PINK_MOSS) {
             return ModBlocks.BUSHY_GRASS.getDefaultState();
         } else if (block == ModBlocks.JUNGLE_MOSS) {
-            return getRandomGrassState(world.rand, ModBlocks.TWISTED_UMBRELLA_MOSS.getDefaultState(),
+            return betterEndForge_1_12_2$getRandomGrassState(world.rand, ModBlocks.TWISTED_UMBRELLA_MOSS.getDefaultState(),
                     ModBlocks.SMALL_JELLYSHROOM.getDefaultState(), ModBlocks.JUNGLE_GRASS.getDefaultState());
         } else if (block == ModBlocks.SANGNUM || block == ModBlocks.MOSSY_DRAGON_BONE || block == ModBlocks.MOSSY_OBSIDIAN) {
-            return getRandomGrassState(world.rand, ModBlocks.GLOBULAGUS.getDefaultState(),
+            return betterEndForge_1_12_2$getRandomGrassState(world.rand, ModBlocks.GLOBULAGUS.getDefaultState(),
                     ModBlocks.CLAWFERN.getDefaultState(), ModBlocks.SMALL_AMARANITA_MUSHROOM.getDefaultState());
         } else if (block == ModBlocks.RUTISCUS) {
-            return getRandomGrassState(world.rand, ModBlocks.AERIDIUM.getDefaultState(),
+            return betterEndForge_1_12_2$getRandomGrassState(world.rand, ModBlocks.AERIDIUM.getDefaultState(),
                     ModBlocks.LAMELLARIUM.getDefaultState(), ModBlocks.BOLUX_MUSHROOM.getDefaultState(), ModBlocks.ORANGO.getDefaultState(), ModBlocks.LUTEBUS.getDefaultState());
         }
         return null;
     }
 
-    private IBlockState getWaterGrassState(World world, BlockPos pos) {
+    @Unique
+    private IBlockState betterEndForge_1_12_2$getWaterGrassState(World world, BlockPos pos) {
         BetterEndBiome biome = ModBiomes.getFromBiome(world.getBiome(pos));
 
         if (world.rand.nextInt(16) == 0) {
@@ -193,36 +206,40 @@ public abstract class BoneMealItemMixin {
         } else if (biome == ModBiomes.SULPHUR_SPRINGS) {
             return world.rand.nextBoolean() ? ModBlocks.CHARNIA_ORANGE.getDefaultState() : ModBlocks.CHARNIA_GREEN.getDefaultState();
         } else if (biome == ModBiomes.UMBRELLA_JUNGLE) {
-            return getRandomGrassState(world.rand, ModBlocks.CHARNIA_CYAN.getDefaultState(),
+            return betterEndForge_1_12_2$getRandomGrassState(world.rand, ModBlocks.CHARNIA_CYAN.getDefaultState(),
                     ModBlocks.CHARNIA_GREEN.getDefaultState(), ModBlocks.CHARNIA_LIGHT_BLUE.getDefaultState());
         } else if (biome == ModBiomes.GLOWING_GRASSLANDS) {
-            return getRandomGrassState(world.rand, ModBlocks.CHARNIA_CYAN.getDefaultState(),
+            return betterEndForge_1_12_2$getRandomGrassState(world.rand, ModBlocks.CHARNIA_CYAN.getDefaultState(),
                     ModBlocks.CHARNIA_GREEN.getDefaultState(), ModBlocks.CHARNIA_LIGHT_BLUE.getDefaultState());
         }
         return ModBlocks.CHARNIA_RED.getDefaultState();
         //return null;
     }
 
-    private IBlockState getRandomGrassState(Random rand, IBlockState... states) {
+    @Unique
+    private IBlockState betterEndForge_1_12_2$getRandomGrassState(Random rand, IBlockState... states) {
         int index = rand.nextInt(states.length);
         return states[index];
     }
 
-    private void shuffle(Random random) {
+    @Unique
+    private void betterEndForge_1_12_2$shuffle(Random random) {
         for (int i = 0; i < 4; i++) {
             int j = random.nextInt(4);
-            EnumFacing d = DIR[i];
-            DIR[i] = DIR[j];
-            DIR[j] = d;
+            EnumFacing d = betterEndForge_1_12_2$DIR[i];
+            betterEndForge_1_12_2$DIR[i] = betterEndForge_1_12_2$DIR[j];
+            betterEndForge_1_12_2$DIR[j] = d;
         }
     }
 
-    private IBlockState getNylium(World world, BlockPos pos) {
-        shuffle(world.rand);
-        for (EnumFacing dir : DIR) {
+    @Unique
+    private IBlockState betterEndForge_1_12_2$getNylium(World world, BlockPos pos) {
+        betterEndForge_1_12_2$shuffle(world.rand);
+        for (EnumFacing dir : betterEndForge_1_12_2$DIR) {
             IBlockState state = world.getBlockState(pos.offset(dir));
-            if (BlockHelper.isEndNylium(state))
+            if (BlockHelper.isEndNylium(state)) {
                 return state;
+            }
         }
         return null;
     }
