+++ src/main/java/mod/beethoven92/betterendforge/mixin/minecraft/ServerLevelMixin.java
@@ -1,14 +1,9 @@
-package mod.beethoven92.betterendforge.mixin;
+package mod.beethoven92.betterendforge.mixin.minecraft;
 
-import mod.beethoven92.betterendforge.BetterEnd;
 import mod.beethoven92.betterendforge.common.init.ModBiomes;
-import mod.beethoven92.betterendforge.common.init.ModBlocks;
 import mod.beethoven92.betterendforge.common.world.generator.GeneratorOptions;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.init.Blocks;
 import net.minecraft.profiler.Profiler;
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraft.world.WorldProvider;
@@ -17,42 +12,43 @@ import net.minecraft.world.storage.ISaveHandler;
 import net.minecraft.world.storage.WorldInfo;
 import net.minecraftforge.fml.common.registry.ForgeRegistries;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(WorldServer.class)
 public abstract class ServerLevelMixin extends World {
-	private static String be_lastWorld = null;
+    @Unique
+    private static String be_lastWorld = null;
 
-	protected ServerLevelMixin(ISaveHandler saveHandlerIn, WorldInfo info, WorldProvider providerIn, Profiler profilerIn, boolean client) {
-		super(saveHandlerIn, info, providerIn, profilerIn, client);
-	}
+    protected ServerLevelMixin(ISaveHandler saveHandlerIn, WorldInfo info, WorldProvider providerIn, Profiler profilerIn, boolean client) {
+        super(saveHandlerIn, info, providerIn, profilerIn, client);
+    }
 
 
-	@Inject(method = "<init>*", at = @At("TAIL"))
-	private void be_onServerWorldInit(MinecraftServer server, ISaveHandler saveHandlerIn, WorldInfo info, int dimensionId, Profiler profilerIn, CallbackInfo ci) {
-		if (be_lastWorld != null && be_lastWorld.equals(saveHandlerIn.loadWorldInfo().getWorldName())) {
-			return;
-		}
-		
-		be_lastWorld = saveHandlerIn.loadWorldInfo().getWorldName();
-		WorldServer world = WorldServer.class.cast(this);
-		ModBiomes.onWorldLoad(world.getSeed(), ForgeRegistries.BIOMES);
-	}
-	
-	@Inject(method = "getSpawnCoordinate", at = @At("HEAD"), cancellable = true)
-	private void be_getSharedSpawnPos(CallbackInfoReturnable<BlockPos> info) {
-		if (GeneratorOptions.changeSpawn()) {
-			if (WorldServer.class.cast(this).provider.getDimension() == 1) {
-				BlockPos pos = GeneratorOptions.getSpawn();
-				info.setReturnValue(pos);
-			}
-		}
-	}
-	
+    @Inject(method = "<init>*", at = @At("TAIL"))
+    private void be_onServerWorldInit(MinecraftServer server, ISaveHandler saveHandlerIn, WorldInfo info, int dimensionId, Profiler profilerIn, CallbackInfo ci) {
+        if (be_lastWorld != null && be_lastWorld.equals(saveHandlerIn.loadWorldInfo().getWorldName())) {
+            return;
+        }
+
+        be_lastWorld = saveHandlerIn.loadWorldInfo().getWorldName();
+        WorldServer world = WorldServer.class.cast(this);
+        ModBiomes.onWorldLoad(world.getSeed(), ForgeRegistries.BIOMES);
+    }
+
+    @Inject(method = "getSpawnCoordinate", at = @At("HEAD"), cancellable = true)
+    private void be_getSharedSpawnPos(CallbackInfoReturnable<BlockPos> info) {
+        if (GeneratorOptions.changeSpawn()) {
+            if (WorldServer.class.cast(this).provider.getDimension() == 1) {
+                BlockPos pos = GeneratorOptions.getSpawn();
+                info.setReturnValue(pos);
+            }
+        }
+    }
+
 //	@Inject(method = "func_241121_a_", at = @At("HEAD"), cancellable = true)
 //	private static void be_createObsidianPlatform(WorldServer serverLevel, CallbackInfo info) {
 //		if (!GeneratorOptions.generateObsidianPlatform()) {
@@ -72,7 +68,7 @@ public abstract class ServerLevelMixin extends World {
 //			info.cancel();
 //		}
 //	}
-	
+
 //	@ModifyArg(method = "tickEnvironment", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/server/ServerWorld;setBlockState(Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/block/BlockState;)Z"))
 //	private IBlockState be_modifyTickState(BlockPos pos, IBlockState state) {
 //		if (state.getBlock()==(Blocks.ICE)) {
