+++ src/main/java/mod/beethoven92/betterendforge/mixin/minecraft/SoundVolumeAccessor.java
@@ -1,12 +1,11 @@
-package mod.beethoven92.betterendforge.mixin;
+package mod.beethoven92.betterendforge.mixin.minecraft;
 
 import net.minecraft.client.audio.PositionedSound;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 @Mixin(PositionedSound.class)
-public interface SoundVolumeAccessor 
-{
-	@Accessor("volume")
-	void setVolume(float volume);
+public interface SoundVolumeAccessor {
+    @Accessor("volume")
+    void setVolume(float volume);
 }
